🐛 negative case

+++ Exercises/501-1000/601-625/616DefuseBomb/defuseBomb.rs
@@ -78,10 +78,11 @@ impl Solution {
       // case negative k
       for i in 0..sz {
         let mut sm = 0;
-        for j in (k..=-1).rev() {
-          let mut idx = i + j as usize;
-          idx = (idx % sz+ sz) % sz; // more complicated than i thought
-          sm+=code[idx];
+        for j in (k..=(-1)).rev() {
+          let mut idx = i as i32 + j as i32; // TIL: usize doesn't support negative numbers.
+          idx = ((idx % sz as i32) + sz as i32) % sz as i32; // more complicated than i thought
+
+          sm+=code[idx as usize];
         }
         res.push(sm);
       }
